requirements.txt
aioboto3==15.5.0
aiobotocore==2.25.1
aiofiles==25.1.0
aiohttp==3.9.5
aioitertools==0.13.0
aiosignal==1.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.14.3
bigcodebench==0.2.0
blinker==1.4
boto3==1.40.61
botocore==1.40.61
bounded-pool-executor==0.0.3
certifi==2023.7.22
cfgv==3.5.0
charset-normalizer==3.1.0
click==8.1.8
comm==0.1.2
contourpy==1.3.3
coverage==7.13.4
cryptography==3.4.8
cycler==0.12.1
datasets==4.7.0
debugpy==1.6.6
decorator==5.1.1
dill==0.4.0
distlib==0.4.0
distro==1.7.0
docker==6.1.3
docstring_parser==0.17.0
evalplus==0.1.5
execnet==2.1.2
executing==1.2.0
filelock==3.25.1
fire==0.7.1
fonttools==4.62.0
frozenlist==1.4.1
fsspec==2025.3.0
gradio_client==2.3.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
huggingface_hub==1.6.0
hydra-core==1.3.2
identify==2.6.17
idna==3.7
ijson==3.5.0
importlib-metadata==6.0.0
iniconfig==2.3.0
inspect_ai==0.3.96
ipykernel==6.21.3
ipython==8.11.0
jedi==0.18.2
jeepney==0.7.1
jiter==0.13.0
jmespath==1.1.0
json-stream==2.4.1
json-stream-rs-tokenizer==0.5.0
jsonlines==4.0.0
jsonpatch==1.33
jsonpath-ng==1.8.0
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.0.3
jupyter_core==5.2.0
keyring==23.5.0
kiwisolver==1.5.0
lark==1.3.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Levenshtein==0.27.3
linkify-it-py==2.1.0
markdown-it-py==4.0.0
matplotlib==3.10.8
matplotlib-inline==0.1.6
mdit-py-plugins==0.5.0
mdurl==0.1.2
mmh3==5.2.1
more-itertools==8.10.0
msgpack==1.1.2
multidict==6.0.5
multipledispatch==1.0.0
multiprocess==0.70.18
nest-asyncio==1.5.6
nest-asyncio2==1.7.2
nodeenv==1.10.0
numpy==1.26.4
oauthlib==3.2.0
omegaconf==2.3.0
openai==2.26.0
packaging==26.0
pandas==3.0.1
parso==0.8.3
pathlib_abc==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==12.1.1
platformdirs==4.2.2
pluggy==1.6.0
pqdm==0.2.0
pre_commit==4.5.1
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.3.0
pyparsing==3.3.2
pyright==1.1.408
pytest==9.0.2
pytest-cov==7.0.0
pytest-xdist==3.8.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==26.0.3
RapidFuzz==3.14.3
referencing==0.37.0
regex==2026.2.28
requests==2.32.2
rich==13.9.4
rpds-py==0.30.0
ruff==0.15.5
s3fs==2025.3.0
s3transfer==0.14.0
scipy==1.17.1
seaborn==0.13.2
SecretStorage==3.3.1
semver==3.0.4
setuptools==82.0.1
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.1
soupsieve==2.8.3
stack-data==0.6.2
tempdir==0.7.1
tenacity==9.1.2
termcolor==3.3.0
textual==6.2.1
tiktoken==0.12.0
tornado==6.4.1
tqdm==4.66.3
traitlets==5.9.0
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
typer==0.23.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uc-micro-py==2.0.0
universal_pathlib==0.3.10
urllib3==1.26.18
virtualenv==20.39.1
wadllib==1.3.6
wcwidth==0.2.6
websocket-client==1.9.0
wget==3.2
wheel==0.46.3
wrapt==1.17.3
xxhash==3.6.0
yarl==1.9.4
z3-solver==4.16.0.0
zipfile-zstd==0.0.4
zipp==3.23.0
zstandard==0.25.0
